requirements.txt
pydantic
outscraper
fastapi
uvicorn
openai
python-dotenv